refactor: how I might think of this app 😀
